Add files via upload
This is a Metals and Mining industry report that I have worked on over the summer.  It contains summaries of news and analyst reports that I read elsewhere on the internet. 
Although I did not directly participated in any of the researches in this file, it still shows my understanding into the industry and my passion for writing.