Add files via upload
This is a Metals and Mining Industry report that I wrote over the summer for the UTEFA investment club at U of T. It contains summaries of major news events and analyst reports I read on the internet. Although I did not directly participated in any of the statistical research in this report, it still showcased my understanding in the Metals and Mining Industry and my ability to compile my writings into a well structured report format!